% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{delta}
\alias{delta}
\title{Generic delta method}
\usage{
delta(g, mean, cov, ses = TRUE)
}
\arguments{
\item{g}{formula (without LHS)}

\item{mean}{named vector (of coefficients)}

\item{cov}{covariance matrix}

\item{ses}{standard errors or (co) variance}
}
\value{
(co) variance of function
}
\description{
Adjusted from \code{msm} package
}
\keyword{internal}
