% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{deltamethod}
\alias{deltamethod}
\title{Delta method}
\usage{
deltamethod(model, formula, robust = TRUE, ses = TRUE)
}
\arguments{
\item{ses}{}
}
\value{

}
\description{
Delta method
}
\examples{
\dontrun{
  deltamethod(model, ~ B_tt + B_cost, ses = F)
  cov <- robvcov(model)
  cov["B_tt", "B_tt"] + cov["B_cost", "B_cost"] + 2 * cov["B_tt", "B_cost"]
}
}
