% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helper.R
\name{estimation_helper}
\alias{estimation_helper}
\title{Helper to generate default availabilites and starting values}
\usage{
estimation_helper(data, utility_script, nDraws, ...)
}
\arguments{
\item{...}{}
}
\description{
Wrapper around \code{mixl::estimate}
}
