% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{fashion_show}
\alias{fashion_show}
\title{A fashion show is a model comparison}
\usage{
fashion_show(models, output_file = NULL, stars = c(0.01, 0.05, 0.1), ...)
}
\arguments{
\item{models}{named list (names are used in table output)}

\item{output_file}{.txt file}

\item{stars}{defaults \code{c(0.01, 0.05, 0.1)}}

\item{...}{passed to \code{texreg::texreg}}
}
\value{
LaTex table output
}
\description{
A fashion show is a model comparison
}
