% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{partworth}
\alias{partworth}
\title{Generate inputs for partworth analysis}
\usage{
partworth(model, na = 9999, file = NULL)
}
\arguments{
\item{model}{output from \code{mixl::estimate}}

\item{na}{NA encodings; defaults to 9999}

\item{file}{full path to file (should end with .xlsx); defaults to \code{NULL}}
}
\description{
Generate inputs for partworth analysis
}
