% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_check.R
\name{stability_check}
\alias{stability_check}
\title{Check stability of model estimates}
\usage{
stability_check(data, utility_script, nDraws, mean = 1, sd = 1)
}
\arguments{
\item{sd}{}
}
\description{
Randomly creates starting values
}
